remove sample snippet and update in next PR.
